backend/requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18
opencv-python-headless>=4.10.0
Pillow>=11.0.0
fonttools[woff]>=4.55.0
brotli>=1.1.0
numpy>=2.0.0
pypotrace>=0.3
